requirements.txt
requests
pandas
gspread
google-auth
google-cloud-secret-manager
